feat(archi): Reorganize structure project
