+++ src/app/app-routing.module.ts
@@ -1,18 +1,17 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuardService } from '@app/guards/auth-guard.service';
 
 const routes: Routes = [
   { path: '', redirectTo: 'log-in', pathMatch: 'full' },
   {
     path: 'sign-up',
-    loadChildren: './auth/sign-up/sign-up.module#SignUpPageModule',
+    loadChildren: '@app/pages/auth/sign-up/sign-up.module#SignUpPageModule',
   },
   {
     path: 'log-in',
-    loadChildren: './auth/log-in/log-in.module#LogInPageModule',
+    loadChildren: '@app/pages/auth/log-in/log-in.module#LogInPageModule',
   },
-  { path: 'home', loadChildren: './tabs/tabs.module#TabsPageModule' },
+  { path: 'home', loadChildren: '@app/pages/tabs/tabs.module#TabsPageModule' },
   { path: '**', redirectTo: 'log-in' },
 ];
