+++ src/app/app.module.ts
@@ -13,8 +13,8 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { JwtModule, JWT_OPTIONS } from '@auth0/angular-jwt';
 import { environment } from '@environments/environment';
-import { TOKEN_KEY } from '@app/shared/service/auth/auth.service';
-import { ErrorInterceptor } from '@app/helpers/error-interceptor/error.interceptor';
+import { TOKEN_KEY } from '@app/core/services/auth/auth.service';
+import { ErrorInterceptor } from '@app/core/interceptors/error.interceptor';
 
 const domainServer = environment.domain_server;
 const domainServerUrn = domainServer.replace(/^http(s?):\/\//i, '');
