+++ src/app/core/guards/auth-guard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
-import { AuthService } from '@app/shared/service/auth/auth.service';
+import { AuthService } from '@app/core/services/auth/auth.service';
 import { Observable } from 'rxjs';
 
 @Injectable({
