+++ src/app/pages/auth/sign-up/sign-up.page.scss
