+++ src/app/pages/tabs/tabs.page.scss
