+++ src/index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <title>Ionic App</title>
+  <title>Memo-riz</title>
 
   <base href="/" />
 
@@ -19,6 +19,7 @@
 </head>
 
 <body>
+  <noscript>Sorry, but the app is not available without JavaScript</noscript>
   <app-root></app-root>
 </body>
